Answered some more of the questions
(also, this is Henry again, git confusion is still being sad.
I forgot to note in commit f0cd04a2ff407ce4d2fb0ce8c25b4346bbe26bf6
that that commit is also me. Also, the reason the timestamps are weird
is that this is getting commited on my pi, which thinks it's in UTC)