Read-ahead and periodic flushing now mostly done. For some reason, some real_time_sleep call blows up when I try to flush in filesys_done. Presumably flushing takes too long and somehting that was sleeping manages to wake up in time to discover that interrupts are off because everything is turning off and freak out. Fixing timer_sleep to block instead of yield might fix this...?
Also, flushing and reading ahead can obviously take a while, so when a test takes long
enough for one of them to trigger, then it suddenly takes a lot more time. It'll
probably be valuable to stick returns at the top of the two periodic functions when
testing other things so they don't add a lot of time to testing

I'll look at fixing timer_sleep in the morning when I'm not dead tired. And then I'll
work on whatever else hasn't been started. Probably making file indexing better if no
one has started that yet